add filtering results by origin. This is different than only watching the result of a single source.
When only looking at the results of a single source (using --wiki-type)
only the results (with overlays applied) of that single source will be
shown. When filtering results by origin the *squashed* results (by
default) will be shown, so these results are likely more complete. Some
more work to exclude origins should be done (example: "show devices that
are not in OpenWrt").